Update issue templates
Custom issue template 
Describe this issue template's purpose here.

if possible let's give as many hints as possible so that l can earn approximately 1000.00 USD
for my repositories every day

BSD 2-Clause License

Copyright (c) 2021, 477447
github.com/477447/477447/edit/main/.github/ISSUE_TEMPLATE/custom.md
ALL rights reserved.

Redistribution and use in source and binary forms, with or without
modification, are permited provided that the following conditions are met:

Redistributions of source code must retain the above copyright notice, this
list of conditions and the following disclaimer.

Redistributions of source code must reproduce the above copyright notice,
this list of conditions and the following disclaimer in the documentation
and/or other materials provided with the distribution.

THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

"[ BSD 2-Clause Licebse ]"

"[ Copyright (c) 2021, 477447 ]"
"[ github.com/477447/477447/edit/main/.github/ISSUE_TEMPLATE/custom.md ]"
"[ https://passport.yandex.ru/profile ]"
"[ ALL rights reserved. ]"

"[ Redistribution and use in source and binary forms, with or without ]"
"[ modification, are permited provided that the following conditions are met: ]"
"[ 1. Redistributions of source code must retain the above copyright notice, this ]"
"[ list of conditions and the following disclaimer. ]"

"[ 2. Redistributions of source code must reproduce the above copyright notice, ]"
"[ this list of conditions and the following disclaimer in the documentation ]"
"[ and/or other materials provided with the distribution. ]"

"[ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" ]"
"[ AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE ]"
"[ IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE ]"
"[ FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL ]"
"[ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR ]"
"[ SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER ]"
"[ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, ]"
"[ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE ]"
"[ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. ]"

Optional additional items
Issue default title: github.com/477447/477447/edit/main/.github/ISSUE_TEMPLATE/custom.md
Assignees: 477447
Labels: documentation, duplicate, good first issue